+++ CHANGELOG.md
@@ -16,6 +16,13 @@
 + Add explicit support for external types (#353, @NathanReb)
 + Add explicit support for new 5.5 let in constructs: `let type`, `let
   external`, `let class` (#354, @NathanReb)
++ Add `--strict` flag to make warnings fatal and parsing stricter
+  (#356, @NathanReb)
++ Add `--check` flag and corresponding check mode which does not print
+  indented version but instead succeed if the input is already correctly
+  indented and fails otherwise. (#357, @NathanReb)
++ Add `?filename` argument to `Nstream` functions to allow properly
+  setting the filename for parsing errors. (#357, @NathanReb)
 
 ## 1.9.0